+++ custom_components/islamic_prayer_times/manifest.json
@@ -2,10 +2,10 @@
   "domain": "islamic_prayer_times",
   "name": "Islamic Prayer Times",
   "documentation": "https://www.home-assistant.io/integrations/islamic_prayer_times",
-  "requirements": ["prayer_times_calculator==0.0.5"],
+  "requirements": ["prayer_times_calculator==0.0.6"],
   "codeowners": ["@engrbm87"],
   "config_flow": true,
   "iot_class": "cloud_polling",
   "loggers": ["prayer_times_calculator"],
-  "version": "2022.6.0"
+  "version": "2022.9.0"
 }
