+++ custom_components/islamic_prayer_times/sensor.py
@@ -1,5 +1,7 @@
 """Platform to retrieve Islamic prayer times information for Home Assistant."""
 
+from datetime import datetime
+
 from homeassistant.components.sensor import (
     SensorDeviceClass,
     SensorEntity,
@@ -10,12 +12,21 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.device_registry import DeviceEntryType
 from homeassistant.helpers.entity import DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
-import homeassistant.util.dt as dt_util
 
-from . import IslamicPrayerDataCoordinator
-from .const import DOMAIN, PRAYER_TIMES_ICON, SENSOR_TYPES
+from .const import DOMAIN, PRAYER_TIMES_ICON
+from .coordinator import IslamicPrayerDataUpdateCoordinator
+
+SENSOR_DESCRIPTIONS: tuple[SensorEntityDescription, ...] = (
+    SensorEntityDescription(key="Fajr", name="Fajr prayer"),
+    SensorEntityDescription(key="Dhuhr", name="Dhuhr prayer"),
+    SensorEntityDescription(key="Asr", name="Asr prayer"),
+    SensorEntityDescription(key="Maghrib", name="Maghrib prayer"),
+    SensorEntityDescription(key="Isha", name="Isha prayer"),
+    SensorEntityDescription(key="Sunrise", name="Sunrise time"),
+    SensorEntityDescription(key="Imsak", name="Imsak time"),
+    SensorEntityDescription(key="Midnight", name="Midnight time"),
+)
 
 
 async def async_setup_entry(
@@ -24,27 +35,26 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up the Islamic prayer times sensor platform."""
-
     coordinator = hass.data[DOMAIN]
 
-    entities = []
-    for description in SENSOR_TYPES:
-        entities.append(IslamicPrayerTimeSensor(coordinator, description))
+    async_add_entities(
+        IslamicPrayerTimeSensor(coordinator, description)
+        for description in SENSOR_DESCRIPTIONS
+    )
 
-    async_add_entities(entities)
 
-
-class IslamicPrayerTimeSensor(CoordinatorEntity, SensorEntity):
+class IslamicPrayerTimeSensor(
+    CoordinatorEntity[IslamicPrayerDataUpdateCoordinator], SensorEntity
+):
     """Representation of an Islamic prayer time sensor."""
 
-    coordinator: IslamicPrayerDataCoordinator
     _attr_device_class = SensorDeviceClass.TIMESTAMP
     _attr_icon = PRAYER_TIMES_ICON
     _attr_should_poll = False
 
     def __init__(
         self,
-        coordinator: IslamicPrayerDataCoordinator,
+        coordinator: IslamicPrayerDataUpdateCoordinator,
         description: SensorEntityDescription,
     ) -> None:
         """Initialize the Islamic prayer time sensor."""
@@ -58,8 +68,6 @@ class IslamicPrayerTimeSensor(CoordinatorEntity, SensorEntity):
         )
 
     @property
-    def native_value(self) -> StateType:
+    def native_value(self) -> datetime:
         """Return the state of the sensor."""
-        return self.coordinator.data[self.entity_description.key].astimezone(
-            dt_util.UTC
-        )
+        return self.coordinator.data[self.entity_description.key]
