+++ custom_components/islamic_prayer_times/translations/en.json
